# docs/swagger.yaml
basePath: /
definitions:
  dto.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  dto.CreateSubscriptionResponse:
    properties:
      id:
        type: string
    type: object
  dto.TotalResponse:
    properties:
      total:
        type: integer
    type: object
  dto.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
    type: object
  models.Subscription:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для управления подписками пользователей.
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Get all subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Subscription'
            type: array
        "500":
          description: Internal Server Error
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription
      parameters:
      - description: Subscription
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateSubscriptionResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      description: Get subscription by UUID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Subscription'
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      description: Calculate total subscription cost for selected period
      parameters:
      - description: User ID
        in: query
        name: user_id
        required: false
        type: string
      - description: Service name
        in: query
        name: service_name
        required: false
        type: string
      - description: Start period (MM-YYYY)
        in: query
        name: from
        required: true
        type: string
      - description: End period (MM-YYYY)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.TotalResponse'
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: Calculate total subscription cost
      tags:
      - subscriptions
swagger: "2.0"